[speech-cmd] reorganize the code
