+++ applications/speech-cmd-analysis/model.py
