+++ applications/speech-cmd-analysis/pipeline.py
